requirements.txt
flask==1.1.2
werkzeug==1.0.1
python-dotenv==0.15.0
google-cloud-logging==2.2.0
discord_interactions==0.2.0
flask-caching==1.10.0
responses==0.12.1
pytest==6.2.2